fix: move v1 create app tests
