+++ server/test/routes/v2/apps.js
@@ -41,39 +41,15 @@ describe('v2/apps', () => {
     })
 
     const sampleApp = {
-        name: 'DHIS2 Sample App',
-        description: 'A very nice sample description',
         appType: 'APP',
-        sourceUrl: 'http://github.com',
         developer: {
-            name: 'Foo Bar',
-            email: 'foobar@dhis2.org',
             organisationId: organisations[0].id,
-        },
-        version: {
-            version: '1.0.0',
-            minDhisVersion: '2.25',
-            maxDhisVersion: '2.33',
-            demoUrl: 'https://www.dhis2.org',
-            channel: 'stable',
         },
     }
 
     const createFormForApp = app => {
         const form = new FormData()
         form.append('app', JSON.stringify(app))
-        form.append(
-            'file',
-            fs.createReadStream(
-                path.join(__dirname, '../../', 'sample-app.zip')
-            )
-        )
-        form.append(
-            'logo',
-            fs.createReadStream(
-                path.join(__dirname, '../../', 'sample-app-logo.png')
-            )
-        )
         return form
     }
 
@@ -92,26 +68,6 @@ describe('v2/apps', () => {
             const receivedPayload = JSON.parse(res.payload)
             expect(receivedPayload).to.include(['id'])
             expect(receivedPayload.id).to.be.string()
-        })
-
-        it('should return 400 bad request if version is not valid', async () => {
-            const badVersionApp = {
-                ...sampleApp,
-                version: {
-                    ...sampleApp.version,
-                    version: 'not a version',
-                },
-            }
-            const form = createFormForApp(badVersionApp)
-            const request = {
-                method: 'POST',
-                url: '/api/v2/apps',
-                headers: form.getHeaders(),
-                payload: await streamToPromise(form),
-            }
-
-            const res = await server.inject(request)
-            expect(res.statusCode).to.equal(400)
         })
     })
 })
